knowledge distilation from text encoder to attr embeder
