+++ configs/data/celebA.yaml
@@ -4,4 +4,4 @@ num_workers: 8
 caption_file: ./data/celebA/captions.json
 image_folder: ./data/celebA/img_align_celeba
 attr_file: ./data/celebA/attrs.json
-keys: ['attr']
+keys: ['image', 'attr']
